requirements.txt
pandas
numpy
Pillow
kagglehub
scipy
tk